Add original APM2 Simulink Blockset release
The APM2 Simulink blockset is a library that allows users to develop
control solutions in Simulink and flash them to hardware without
requiring extensive experience. It was developed during a graduate level
Guidance, Navigation, and Control course at Embry-Riddle Aeronautical
University by a team of industry professionals and has been downloaded
by thousands of individuals and organizations around the globe.

Details regarding the blockset were published in an American Institute
of Aeronautics and Astronautics paper (AIAA 2013-5009) and presented at
the Guidance, Navigation, and Control conference in Boston, MA in
2013. The original release was published to the Mathworks file
exchange; unfortunately, the project has outgrown this type of
distribution model hence the move to github.

This commit is non functional in that it simply captures the project as
it currently exists. It is intended to be a starting point for
integration into the larger ardupilot project on which much of the
Mathworks release was built. This effort is to synchronize these two
entities.

For more information, please visit www.mathworks.com and search for
APM2.